Update JS code
Added a link to download the Cover art
Renamed the files to include the CardTitle and a 3-digit padded track number
Added a metadata text field that pulls lots of non-identifying information from the card so it can be stored in a text file. I couldn't get the JS to create a text file for downloading but I left that code in there in case someone else can fix it.